refactor(subs): share subscription update hook
Extract useUpdateSubscriptionForm holding the mutation, toast and form
lifecycle shared by the per-section update hooks. useUpdateSubscription
Information and useUpdateSubscriptionInvoicingPayments become thin
delegators supplying only buildInput. Behavior-preserving; clears the
SonarCloud new-code duplication gate.
